Moves createWindow function out of appStore.js
Pre-work for #10458 tests

Auditors: @bbondy @bsclifton

Test Plan:
- covered by automated tests
